+++ src/Mono.Android/Microsoft.Android.Runtime/ManagedTypeManager.cs
@@ -59,16 +59,75 @@ class ManagedTypeManager : JniRuntime.JniTypeManager {
 		return MakeGenericType (suffixDefinition, arguments);
 	}
 
+	// NOTE: suppressions below also in `src/Mono.Android/Android.Runtime/AndroidRuntime.cs`
+	[UnconditionalSuppressMessage ("Trimming", "IL2057", Justification = "Type.GetType() can never statically know the string value parsed from parameter 'methods'.")]
+	[UnconditionalSuppressMessage ("Trimming", "IL2067", Justification = "Delegate.CreateDelegate() can never statically know the string value parsed from parameter 'methods'.")]
+	[UnconditionalSuppressMessage ("Trimming", "IL2072", Justification = "Delegate.CreateDelegate() can never statically know the string value parsed from parameter 'methods'.")]
 	public override void RegisterNativeMembers (
 			JniType nativeClass,
 			[DynamicallyAccessedMembers (MethodsAndPrivateNested)]
 			Type type,
 			ReadOnlySpan<char> methods)
 	{
-		if (!NativeMethodRegistrar.TryRegisterNativeMembers (nativeClass, type, methods)) {
+		if (methods.IsEmpty) {
 			base.RegisterNativeMembers (nativeClass, type, methods);
+			return;
+		}
+
+		int methodCount = CountMethods (methods);
+		if (methodCount < 1) {
+			base.RegisterNativeMembers (nativeClass, type, methods);
+			return;
+		}
+
+		JniNativeMethodRegistration [] natives = new JniNativeMethodRegistration [methodCount];
+		int nativesIndex = 0;
+
+		ReadOnlySpan<char> methodsSpan = methods;
+		bool needToRegisterNatives = false;
+
+		while (!methodsSpan.IsEmpty) {
+			int newLineIndex = methodsSpan.IndexOf ('\n');
+
+			ReadOnlySpan<char> methodLine = methodsSpan.Slice (0, newLineIndex != -1 ? newLineIndex : methodsSpan.Length);
+			if (!methodLine.IsEmpty) {
+				SplitMethodLine (methodLine,
+					out ReadOnlySpan<char> name,
+					out ReadOnlySpan<char> signature,
+					out ReadOnlySpan<char> callbackString,
+					out ReadOnlySpan<char> callbackDeclaringTypeString);
+
+				Delegate? callback = null;
+				if (callbackString.SequenceEqual ("__export__")) {
+					throw new InvalidOperationException (FormattableString.Invariant ($"Methods such as {callbackString.ToString ()} are not implemented!"));
+				} else {
+					Type callbackDeclaringType = type;
+					if (!callbackDeclaringTypeString.IsEmpty) {
+						callbackDeclaringType = Type.GetType (callbackDeclaringTypeString.ToString (), throwOnError: true)!;
+					}
+					while (callbackDeclaringType.ContainsGenericParameters) {
+						callbackDeclaringType = callbackDeclaringType.BaseType!;
+					}
+
+					GetCallbackHandler connector = (GetCallbackHandler) Delegate.CreateDelegate (typeof (GetCallbackHandler),
+						callbackDeclaringType, callbackString.ToString ());
+					callback = connector ();
+				}
+
+				if (callback != null) {
+					needToRegisterNatives = true;
+					natives [nativesIndex++] = new JniNativeMethodRegistration (name.ToString (), signature.ToString (), callback);
+				}
+			}
+
+			methodsSpan = newLineIndex != -1 ? methodsSpan.Slice (newLineIndex + 1) : default;
+		}
+
+		if (needToRegisterNatives) {
+			JniEnvironment.Types.RegisterNatives (nativeClass.PeerReference, natives, nativesIndex);
 		}
 	}
+
 
 	protected override IEnumerable<Type> GetTypesForSimpleReference (string jniSimpleReference)
 	{
@@ -97,4 +156,39 @@ class ManagedTypeManager : JniRuntime.JniTypeManager {
 	{
 		return JniRemappingLookup.GetStaticMethodFallbackTypes (jniSimpleReference, useReplacementTypes: false);
 	}
+
+	static int CountMethods (ReadOnlySpan<char> methodsSpan)
+	{
+		int count = 0;
+		while (!methodsSpan.IsEmpty) {
+			count++;
+
+			int newLineIndex = methodsSpan.IndexOf ('\n');
+			methodsSpan = newLineIndex != -1 ? methodsSpan.Slice (newLineIndex + 1) : default;
+		}
+		return count;
+	}
+
+	static void SplitMethodLine (
+		ReadOnlySpan<char> methodLine,
+		out ReadOnlySpan<char> name,
+		out ReadOnlySpan<char> signature,
+		out ReadOnlySpan<char> callback,
+		out ReadOnlySpan<char> callbackDeclaringType)
+	{
+		int colonIndex = methodLine.IndexOf (':');
+		name = methodLine.Slice (0, colonIndex);
+		methodLine = methodLine.Slice (colonIndex + 1);
+
+		colonIndex = methodLine.IndexOf (':');
+		signature = methodLine.Slice (0, colonIndex);
+		methodLine = methodLine.Slice (colonIndex + 1);
+
+		colonIndex = methodLine.IndexOf (':');
+		callback = methodLine.Slice (0, colonIndex != -1 ? colonIndex : methodLine.Length);
+
+		callbackDeclaringType = colonIndex != -1 ? methodLine.Slice (colonIndex + 1) : default;
+	}
+
+	delegate Delegate GetCallbackHandler ();
 }
