--- src/Mono.Android/Microsoft.Android.Runtime/NativeMethodRegistrar.cs
@@ -1,140 +1,0 @@
-#nullable enable
-
-using System;
-using System.Diagnostics.CodeAnalysis;
-using Java.Interop;
-
-namespace Microsoft.Android.Runtime;
-
-/// <summary>
-/// Shared implementation of the legacy, reflection-based JNI native method registration.
-/// Parses the <c>methods</c> metadata string (produced for Java Callable Wrappers and
-/// <c>net/dot/jni/ManagedPeer.registerNativeMembers</c>) and registers the native callbacks
-/// via <see cref="JniEnvironment.Types.RegisterNatives"/>.
-///
-/// This is the "slow" path: it relies on <see cref="Delegate.CreateDelegate"/> and
-/// <see cref="Type.GetType(string, bool)"/>, so it is not trim-friendly. It is used directly by
-/// <see cref="ManagedTypeManager"/> and, behind <see cref="RuntimeFeature.LegacyJniRegistration"/>,
-/// by <see cref="TrimmableTypeMapTypeManager"/>.
-/// </summary>
-static class NativeMethodRegistrar
-{
-	const DynamicallyAccessedMemberTypes MethodsAndPrivateNested =
-		DynamicallyAccessedMemberTypes.PublicMethods |
-		DynamicallyAccessedMemberTypes.NonPublicMethods |
-		DynamicallyAccessedMemberTypes.NonPublicNestedTypes;
-
-	// NOTE: suppressions below also in `src/Mono.Android/Android.Runtime/AndroidRuntime.cs`
-	// and historically in `ManagedTypeManager`.
-	/// <summary>
-	/// Registers the native members described by <paramref name="methods"/>.
-	/// </summary>
-	/// <returns>
-	/// <see langword="true"/> if the <paramref name="methods"/> string was parsed (whether or not
-	/// any natives were registered); <see langword="false"/> if it was empty / contained no method
-	/// lines, in which case the caller should fall back to
-	/// <c>base.RegisterNativeMembers</c> (the marshal-methods path).
-	/// </returns>
-	[UnconditionalSuppressMessage ("Trimming", "IL2057", Justification = "Type.GetType() can never statically know the string value parsed from parameter 'methods'.")]
-	[UnconditionalSuppressMessage ("Trimming", "IL2067", Justification = "Delegate.CreateDelegate() can never statically know the string value parsed from parameter 'methods'.")]
-	[UnconditionalSuppressMessage ("Trimming", "IL2072", Justification = "Delegate.CreateDelegate() can never statically know the string value parsed from parameter 'methods'.")]
-	public static bool TryRegisterNativeMembers (
-			JniType nativeClass,
-			[DynamicallyAccessedMembers (MethodsAndPrivateNested)]
-			Type type,
-			ReadOnlySpan<char> methods)
-	{
-		if (methods.IsEmpty) {
-			return false;
-		}
-
-		int methodCount = CountMethods (methods);
-		if (methodCount < 1) {
-			return false;
-		}
-
-		JniNativeMethodRegistration [] natives = new JniNativeMethodRegistration [methodCount];
-		int nativesIndex = 0;
-
-		ReadOnlySpan<char> methodsSpan = methods;
-		bool needToRegisterNatives = false;
-
-		while (!methodsSpan.IsEmpty) {
-			int newLineIndex = methodsSpan.IndexOf ('\n');
-
-			ReadOnlySpan<char> methodLine = methodsSpan.Slice (0, newLineIndex != -1 ? newLineIndex : methodsSpan.Length);
-			if (!methodLine.IsEmpty) {
-				SplitMethodLine (methodLine,
-					out ReadOnlySpan<char> name,
-					out ReadOnlySpan<char> signature,
-					out ReadOnlySpan<char> callbackString,
-					out ReadOnlySpan<char> callbackDeclaringTypeString);
-
-				Delegate? callback = null;
-				if (callbackString.SequenceEqual ("__export__")) {
-					throw new InvalidOperationException (FormattableString.Invariant ($"Methods such as {callbackString.ToString ()} are not implemented!"));
-				} else {
-					Type callbackDeclaringType = type;
-					if (!callbackDeclaringTypeString.IsEmpty) {
-						callbackDeclaringType = Type.GetType (callbackDeclaringTypeString.ToString (), throwOnError: true)!;
-					}
-					while (callbackDeclaringType.ContainsGenericParameters) {
-						callbackDeclaringType = callbackDeclaringType.BaseType!;
-					}
-
-					GetCallbackHandler connector = (GetCallbackHandler) Delegate.CreateDelegate (typeof (GetCallbackHandler),
-						callbackDeclaringType, callbackString.ToString ());
-					callback = connector ();
-				}
-
-				if (callback != null) {
-					needToRegisterNatives = true;
-					natives [nativesIndex++] = new JniNativeMethodRegistration (name.ToString (), signature.ToString (), callback);
-				}
-			}
-
-			methodsSpan = newLineIndex != -1 ? methodsSpan.Slice (newLineIndex + 1) : default;
-		}
-
-		if (needToRegisterNatives) {
-			JniEnvironment.Types.RegisterNatives (nativeClass.PeerReference, natives, nativesIndex);
-		}
-
-		return true;
-	}
-
-	static int CountMethods (ReadOnlySpan<char> methodsSpan)
-	{
-		int count = 0;
-		while (!methodsSpan.IsEmpty) {
-			count++;
-
-			int newLineIndex = methodsSpan.IndexOf ('\n');
-			methodsSpan = newLineIndex != -1 ? methodsSpan.Slice (newLineIndex + 1) : default;
-		}
-		return count;
-	}
-
-	static void SplitMethodLine (
-		ReadOnlySpan<char> methodLine,
-		out ReadOnlySpan<char> name,
-		out ReadOnlySpan<char> signature,
-		out ReadOnlySpan<char> callback,
-		out ReadOnlySpan<char> callbackDeclaringType)
-	{
-		int colonIndex = methodLine.IndexOf (':');
-		name = methodLine.Slice (0, colonIndex);
-		methodLine = methodLine.Slice (colonIndex + 1);
-
-		colonIndex = methodLine.IndexOf (':');
-		signature = methodLine.Slice (0, colonIndex);
-		methodLine = methodLine.Slice (colonIndex + 1);
-
-		colonIndex = methodLine.IndexOf (':');
-		callback = methodLine.Slice (0, colonIndex != -1 ? colonIndex : methodLine.Length);
-
-		callbackDeclaringType = colonIndex != -1 ? methodLine.Slice (colonIndex + 1) : default;
-	}
-
-	delegate Delegate GetCallbackHandler ();
-}
